// src/types/User.ts
export type User = {
  DN: string;
  Uid: string;
  Cn: string;
  Sn: string;
  Gn: string;
  Mail: string;
  UidNumber: number;
  StudentGender: string;
  StudentBirthday: string;
  TelephoneNumber: string;
  StudentColleage: string;
  StudentFirstMajor: string;
  StudentMajor: string;
  StudentEnrolled: boolean;
  StudentGraduated: boolean;
};